merge_baidu_2024_12_10 compile 8
